dockstrings and refactor a little
